A start at python3 support
This should be 100% compatible with python2, and should not affect the existing ability to run. It is mostly changing print statements to use the print function.

When attempting to package with python3, it fails to find a version of sklearn that is compatible with python3 *and* has the sklearn.cross_validation module (which went away).